fix(go): Make SHOW WAREHOUSES test less flaky (#1494)
Minor fix to a test that has been breaking each time Snowflake added a
new field to the result of a `SHOW WAREHOUSES` query.

Previously we asserted the number of columns, but this number has been
changing. Instead of completely removing this part of the test, we can
assert the columns we know currently should be present and expect that
Snowflake won't make breaking changes to the schema by removing columns.
The specific contents of the schema are not relevant to the original
purpose of the test anyway, so it should be fine if the actual schema
continues to grow beyond what is asserted here.