fix(go): Records dropped on ingestion when empty batch is present (#1866)
Reproduces and fixes: #1847 

Parquet files with empty row groups are valid per the spec, but
Snowflake does not currently handle them properly. To mitigate this we
buffer writes to the parquet file so that a row group is not written
until some amount of data has been received.

The CheckedAllocator was enabled for all tests as part of this fix,
which detected a leak in the BufferWriter that was fixed in:
[https://github.com/apache/arrow/pull/41698](https://github.com/apache/arrow/pull/41698).

There was an unrelated test failure that surfaced once the
CheckedAllocator was enabled which had to do with casting decimals of
certain precision. The fix is included in this PR as well.