fix(go): use one session for connection (#2494)
Instead of creating a separate connection for simple queries like SELECT
COUNT and the metadata queries, we'll keep using the single connection
to maintain the existing session throughout. This should avoid the
problem mentioned in #2128 where a new session cancels the remaining
COPY INTO queries. It also simplifies some of the work on metadata
queries so that we don't need to explicitly propagate catalog, schema
and database names between connections.

Fixes #2517